Fix API docs deploy so that a PR is not required (#15011)
Fixes this
[issue](https://github.com/microsoft/onnxruntime/actions/runs/4387534694/jobs/7682945415#step:12:534)
and removes the extra PR step in the workflow.

Also logs the commit of the main branch that the docs were generated
from to a file called version.txt at the root of the API docs tree.

Tested for Java API docs and results staged here:
https://natke.github.io/onnxruntime/docs/api/java/index.html

If approved, I can migrate all of the other API docs generation
workflows to use this scheme.