Enable LeakyRelu latest and refactor fast_gelu_fusion to enable the script (#15003)
### Description
Enable LeakyRelu latest since the last version differs only in type
support.
Refactor `fast_gelu_fusion` to enable the script, because our script is
unable to
check if any of the optimizers are outdated and no longer in effect.

### Motivation and Context
We do not want to loose performance.

Next step is to file improvements issues if any are required.