Handle unused function inputs (#15130)
### Description

Fix issue relating to unused inputs of model-local functions. ORT
creates a schema for all such functions. The creation of this schema
does not handle unused function-inputs. The schema-creation relies on
the use of the function-inputs to infer type-constraints for the input,
and the code ends up creating an erroneous input-descriptor when there
is no use of the function-input.

The fix is to create an input with the given name, with a
type-constraint that allows all types.

### Motivation and Context
Fix https://github.com/microsoft/onnxruntime/issues/15046

Fix https://github.com/microsoft/onnx-script/issues/524

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>
Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>
Co-authored-by: Scott McKay <skottmckay@gmail.com>