Make internal nhwc schema registrations complete (#15278)
### Description
<!-- Describe your changes. -->
Add all the ONNX layout sensitive ops from opset 11 on. 
Make list in transpose optimizer consistent.

### Motivation and Context
<!-- - Why is this change required? What problem does it solve?
- If it fixes an open issue, please link to the issue here. -->
When we run L1 optimizers after layout transform in a full build it
needs a schema for any layout sensitive ops that get converted to the
internal domain. Previously we did not run L1, so we got away without
having schemas unless the EP used a static kernel for the nhwc version
of the op.