Exposing new TRT build options (#15089)
### Description

This will add a few TRT options, some of them are only available on TRT
8.6:
- heuristics
- sparsity
- optimization level (8.6 only)
- auxiliary stream (8.6 only)
- tactic source selection

I am no sure yet which tests is should add for these options. As those
are mostly simple TRT flags i am not sure to what level i should test.
For heuristics something similar to
https://github.com/gedoensmax/onnxruntime/blob/44dda08b51e98bd86f6eafd09b7d00c3680c78a8/onnxruntime/test/providers/tensorrt/tensorrt_basic_test.cc#L510-L538
should be possible for, but for all other essentially we would only be
testing if there is a crash or not if the option is set.
Also if i forgot some option that would be good to have feel free to
speak up !