[QNN EP] Fix shape inference of NHWC Resize (#15477)
### Description
Adds schema for NHWC Resize that uses the default ONNX type/shape
inferencing.


### Motivation and Context
The QNN EP requires the Resize operator to be NHWC. Currently, the
Resize operator fails type and shape inference because the current
schema changes the input to NCHW, but the `scales` and `sizes` inputs
remain in NHWC.

This PR adds a schema for NHWC Resize that allows it to use the default
ONNX type/shape inference while still remaining in the internal NHWC
domain.